test.requirements.txt
requests
pytest==3.6.0
pyhamcrest==1.9.0
